+++ src/index.ts
@@ -1,0 +1,4 @@
+// Types definitions export.
+export * from './models';
+// Strategy default and named export.
+export { default, Strategy } from './strategy';