+++ src/models/twitterError.ts
@@ -1,0 +1,12 @@
+export interface TwitterErrorResponse {
+  data: TwitterError;
+}
+
+export interface TwitterError {
+  errors: [
+    {
+      message: string;
+      code: string;
+    }
+  ];
+}